Results page renders with submitted script
