+++ src/components/app.jsx
@@ -8,15 +8,25 @@ import Results from "./results.jsx";
 class App extends React.Component {
   constructor(props) {
     super(props);
+    this.state = {
+      script: undefined
+    }
+    this.setScript = this.setScript.bind(this);
+  }
+
+  setScript(script) {
+    this.setState({
+      script: script
+    });
   }
 
   render () {
     return (
       <div>
-        <Route exact path='/' component={ Upload } />
+        <Route exact path='/' render={() => <Upload setscript={this.setScript} />} />
         <Route path='/speech' component={ Speech } />
         <Route path='/landing' component={ Landing } />
-        <Route path='/results' component= { Results } />
+        <Route path='/results' render={() => <Results script={this.state.script} />} />
       </div>
     )
   }
